Configured Content-Security-Policy.
In a move to migrate this to a single-file distribution, a necessary
component was to add CSP to this application. We've now identified
all the locations from which we load web-based sources. A future
version may move all of the dependencies directly into the
application.

Also removed a now-unused build script.