Added a new File-IO component area.
The file_io folder is going to be the new file format component. It is similar in
implementation to the /src/collections area, but that area didn't start with the
ability to write in mind, and it just was not working out.

This implementation introduces mocha and chai, and includes a number of test cases
for the Network Console-native file format.