Targeted perf improvements (#23)
* Targeted perf improvements

This change implements targeted perf improvements when the number of messages in the WebSocketView is high. This is
because React would render all the WebSocketMessage components beneath.

Additionally, to improve testability and profiling, this implements a pretty lame "real WebSocket" connection in the
WebApplicationHost. However, it's pretty unusable, unless someone adds a `connect-src:` declaration to the CSP
definition in the index.html file.

* Fixed WS mocks