Final round of changes for the file format feature.
This set of changes fixes a variety of small bugs related to the new
file-IO component. The primary problem involved here was that "Save As"
didn't work for imported requests, and it had to be updated via the
protocol since the frontend had made the assumption that it would be
one-tab-per-request previously.