Added a new File-IO component area. (#14)
* Added a new File-IO component area.

The file_io folder is going to be the new file format component. It is similar in
implementation to the /src/collections area, but that area didn't start with the
ability to write in mind, and it just was not working out.

This implementation introduces mocha and chai, and includes a number of test cases
for the Network Console-native file format.

* Added Authorization proxy and handling

* Started Postman v2.1 save migration

* Started Postman tests

* Added Postman v2.1 read/write capability

This change duplicates the Network Console-native file format
API shape to Postman files, along with the same set of test cases.

* Adds basic OpenAPI v2 support to the new file_io module.

OpenAPI was already supported (and this support more or less migrates directly) from the collections/ module. This approach implements the new API shape.

* Added ability to convert between file types.

This addition discovered a number of bugs in file serialization which have been corrected.

* Added test to convert OpenAPI to Native

* Enhanced Postman-v2.1 URL decomposition support

This change adds the ability to parse and decompose complex URLs
when loading Postman collection files.

It also deprecates the existing Collection parsing code.

Factored out the simple BidiMap from the collections, using a specialization
instead that supports the specific use cases needed for tracking ID-Index
relationships in the collection adapters.

Added a bunch of comments to interfaces where appropriate.

Added tests for Lazy and the ID-Index map.

* Adds read/write/convert support for Environment files

This supports Postman files (mapping to the version currently
exported by Postman, which we're calling "2.1" schema for now,
although the environment files don't define a schema), Network
Console-native file format, and OpenAPI.

The Network Console Environment (*.ncenv.json) format supports
multiple environments per file, and this is reflected in the
overall API shape.

OpenAPI version 2 only supports schemes, a single Base URI (server), and a single base path. In practice, the default environment created by OpenAPI v2 is "http" and "https", if they're defined at all.

This set of changes fixes a variety of small bugs related to the new
file-IO component. The primary problem involved here was that "Save As"
didn't work for imported requests, and it had to be updated via the
protocol since the frontend had made the assumption that it would be
one-tab-per-request previously.

* Version bump to 0.10.0-preview