Enhanced localizability and UX.
This change fixes some localizability defects in EditorGrid along with
transitioning more EditorGrid controls to FAST-React from FluentUI.

Because we made localizability changes, we also made some UX fixes that
were requested by the UX team. Specifically we repositioned the
"response pending" and "request not sent" views, and modified the
top-level 'density' setting for the design system.

This also fixes the `GridTextOrFileKey` control which is the "Key"
column of the EditorGrid in which a grid can choose to go into file or
text mode.