Localizability support and removal of @fluentui dependency (#34)
* Removed deprecated APIs

This change removes the APIs that were deprecated in v0.10.0-preview and
outlines the plan for the rest of 0.11.0-preview.

* Added loc API stubs to network-console-shared

* Initial set of localizability work.

Implemented stubs in NC-Shared for localization. Made the following
components localizable:
 - HttpVerbPicker
 - ErrorBoundary
 - ViewSelect

Additionally implemented helper components.

* Made several components localizable:

 - AddressBar
 - Authorization (root)
 - Authorization - Basic
 - Authorization - Bearer
 - CalculatedAddress
 - ErrorBoundary
 - HttpVerbPicker
 - ViewSelect
 - BoolCheck

Also refactored several components to reduce dependency on Fluent UI.

* Localized the EditorGrid and related components

* More localizability.

Here we localized and fixed up some UX in the address bar and editor
grid, plus we load the English dictionary by default in the frontend.

* Made almost all of the Response Viewer localizable.

* Started migrating the EditorGrid to FAST from Fluent.

The TextField types are close but not quite the same. This work
starts the effort of migrating EditorGrid off of the @fluentui
package.

* Enhanced localizability and UX.

This change fixes some localizability defects in EditorGrid along with
transitioning more EditorGrid controls to FAST-React from FluentUI.

Because we made localizability changes, we also made some UX fixes that
were requested by the UX team. Specifically we repositioned the
"response pending" and "request not sent" views, and modified the
top-level 'density' setting for the design system.

This also fixes the `GridTextOrFileKey` control which is the "Key"
column of the EditorGrid in which a grid can choose to go into file or
text mode.

* Added the pseudoloc generation script and pseudoloc package.

Added "locale" to the frontend Store, hooked up the app's
LocalizationProvider to that, and created the necessary hookup to load
localization via a message or otherwise.

Added the capability to generate a psuedoloc string, and the
web-application-host implementation loads it by default.

* Localized more controls

This change makes more controls localizable:

 - Authorization selector
 - Request Body pivot
 - Request Body mode selector
 - "None" body type message

* More localization and related support.

This change adds more localized components, including:

 - All modals
 - The Fetch configuration screen
 - The Request editor

In addition, it fixes several Fluent->FAST-related bugs that have
cropped up. It includes the transition of the primary modal dialog from
Fluent to FAST.

* Ported the Fetch radiogroup control from Fluent to FAST

This change ports the 'Fetch' configuration radio button group option
sets to FAST-React controls. This greatly simplifies the Fetch
presentation screen.

It also fixes a bug with the modal presentation of the authorization
editor with respect to tabbing into radios.

* Removed the balance of @fluentui dependencies.

This change removes the balance of the Fluent UI dependencies. Much of
the work had already been done, but we had a few remaining cases in
which it was used, primarily around icons.

* Code review feedback from pull request.

Additionally: addressed #30, by allowing the host to include
authorization parameters are part of the "set collections" information,
rather than as part of the "load request" message.

* Added ability to cancel an in-flight request, #32.

* Bumped version to 0.11.0-preview.

* Commented out delay used for testing in Web Application Host